+++ apps/decodex-app/Sources/DecodexApp/OperatorLaneReadoutData.swift
@@ -1,0 +1,223 @@
+import Foundation
+
+extension OperatorLanePopoverView {
+	var activity: OperatorChildAgentActivity? {
+		run.childAgentActivity
+	}
+
+	var currentSummary: String {
+		if run.processAlive == false, run.hasFreshExecution == false {
+			if let idle = formatActivityDuration(run.inactiveDurationSeconds) {
+				return "Stopped · idle \(idle)"
+			}
+
+			return "Stopped"
+		}
+
+		guard let activity else {
+			return "Waiting for child activity"
+		}
+
+		let label = panelTrimmed(activity.currentDetail)
+			?? panelTrimmed(activity.currentBucket).map(rawPanelToken)
+			?? "Active"
+		if let elapsed = formatActivityDuration(activity.currentElapsedSeconds(at: currentTime)) {
+			return "\(rawPanelToken(label)) · \(elapsed)"
+		}
+
+		return rawPanelToken(label)
+	}
+
+	var projectTitle: String? {
+		panelTrimmed(run.projectDisplayName) ?? panelTrimmed(run.projectID)
+	}
+
+	var statusReadoutItems: [OperatorLaneReadoutItem] {
+		var items = [OperatorLaneReadoutItem]()
+
+		if let runPhase = panelTrimmed(run.runPhase ?? run.phase) {
+			items.append(
+				OperatorLaneReadoutItem(label: "run phase", value: rawPanelToken(runPhase))
+			)
+		}
+
+		return items
+	}
+
+	var fallbackRunReadoutItems: [OperatorLaneReadoutItem] {
+		guard let activity else {
+			return []
+		}
+
+		var items = [
+			OperatorLaneReadoutItem(
+				label: "wall",
+				value: formatActivityDuration(activity.wallSeconds(at: currentTime)) ?? "0s"
+			),
+			OperatorLaneReadoutItem(
+				label: "events",
+				value: formatCompactCount(activity.eventCount)
+			),
+			OperatorLaneReadoutItem(
+				label: "input",
+				value: "\(formatCompactCount(activity.inputTokensCumulative)) tok"
+			),
+			OperatorLaneReadoutItem(
+				label: "output",
+				value: "\(formatCompactCount(activity.outputTokensCumulative)) tok"
+			),
+			OperatorLaneReadoutItem(
+				label: "tools",
+				value: formatCompactCount(activity.toolCallCount)
+			),
+		]
+
+		if let largestOutput = activity.largestToolOutputBytes, largestOutput > 0 {
+			items.append(
+				OperatorLaneReadoutItem(
+					label: "max output",
+					value: formatLargestOutput(bytes: largestOutput)
+				)
+			)
+		}
+
+		return items
+	}
+
+	var modelProgress: OperatorModelProgressReadout? {
+		operatorModelProgressReadout(for: run, currentTime: currentTime)
+	}
+
+	var detailBuckets: [OperatorChildAgentBucket] {
+		guard run.lifecycleMetrics == nil else {
+			return []
+		}
+
+		return orderedBuckets.filter { bucket in
+			bucket.name.caseInsensitiveCompare("Model") != .orderedSame
+				&& detailBucketIsVisible(bucket)
+				&& bucketReadoutItems(bucket).isEmpty == false
+		}
+	}
+
+	func detailBucketIsVisible(_ bucket: OperatorChildAgentBucket) -> Bool {
+		let normalizedName = bucket.name.lowercased()
+
+		return normalizedName.contains("protocol") || normalizedName.contains("tracker")
+	}
+
+	var orderedBuckets: [OperatorChildAgentBucket] {
+		bucketRows.sorted { left, right in
+			let leftPriority = bucketPriority(left.name)
+			let rightPriority = bucketPriority(right.name)
+			if leftPriority != rightPriority {
+				return leftPriority < rightPriority
+			}
+			let leftWallSeconds = bucketWallSeconds(left)
+			let rightWallSeconds = bucketWallSeconds(right)
+			if leftWallSeconds != rightWallSeconds {
+				return leftWallSeconds > rightWallSeconds
+			}
+			if left.eventCount != right.eventCount {
+				return left.eventCount > right.eventCount
+			}
+
+			return left.name < right.name
+		}
+	}
+
+	var bucketRows: [OperatorChildAgentBucket] {
+		activity?.buckets ?? []
+	}
+
+	func bucketReadoutItems(_ bucket: OperatorChildAgentBucket) -> [OperatorLaneReadoutItem] {
+		let normalizedName = bucket.name.lowercased()
+		var items = [OperatorLaneReadoutItem]()
+
+		let wallSeconds = bucketWallSeconds(bucket)
+
+		if normalizedName.contains("tracker"), wallSeconds > 0 {
+			items.append(
+				OperatorLaneReadoutItem(
+					label: "wall",
+					value: formatActivityDuration(wallSeconds) ?? "0s"
+				)
+			)
+		}
+		if bucket.eventCount > 0 {
+			items.append(OperatorLaneReadoutItem(label: "events", value: formatCompactCount(bucket.eventCount)))
+		}
+		if normalizedName.contains("protocol") {
+			if bucket.inputTokens > 0 {
+				items.append(
+					OperatorLaneReadoutItem(
+						label: "input",
+						value: "\(formatCompactCount(bucket.inputTokens)) tok"
+					)
+				)
+			}
+			if bucket.outputTokens > 0 {
+				items.append(
+					OperatorLaneReadoutItem(
+						label: "output",
+						value: "\(formatCompactCount(bucket.outputTokens)) tok"
+					)
+				)
+			}
+		} else {
+			if bucket.toolCallCount > 0 {
+				items.append(OperatorLaneReadoutItem(label: "tools", value: formatCompactCount(bucket.toolCallCount)))
+			}
+			if bucket.outputBytes > 0 {
+				items.append(OperatorLaneReadoutItem(label: "output bytes", value: formatCompactBytes(bucket.outputBytes)))
+			}
+			if normalizedName.contains("tracker") == false {
+				if bucket.inputTokens > 0 {
+					items.append(
+						OperatorLaneReadoutItem(
+							label: "input",
+							value: "\(formatCompactCount(bucket.inputTokens)) tok"
+						)
+					)
+				}
+				if bucket.outputTokens > 0 {
+					items.append(
+						OperatorLaneReadoutItem(
+							label: "output",
+							value: "\(formatCompactCount(bucket.outputTokens)) tok"
+						)
+					)
+				}
+			}
+		}
+
+		return items
+	}
+
+	func bucketPriority(_ name: String) -> Int {
+		let normalizedName = name.lowercased()
+		if normalizedName.contains("model") {
+			return 0
+		}
+		if normalizedName.contains("protocol") {
+			return 1
+		}
+		if normalizedName.contains("tracker") {
+			return 2
+		}
+
+		return 10
+	}
+
+	func bucketWallSeconds(_ bucket: OperatorChildAgentBucket) -> Int {
+		activity?.wallSeconds(for: bucket, at: currentTime) ?? bucket.wallSeconds
+	}
+
+	func formatLargestOutput(bytes: Int?) -> String {
+		guard let bytes, bytes > 0 else {
+			return "-"
+		}
+
+		return formatCompactBytes(bytes)
+	}
+}
